+++ nativelink-store/src/redis_store.rs
@@ -34,8 +34,8 @@ use uuid::Uuid;
 use crate::cas_utils::is_zero_digest;
 
 // TODO(caass): These (and other settings) should be made configurable via nativelink-config.
-pub const READ_CHUNK_SIZE: usize = 64 * 1024;
-const CONNECTION_POOL_SIZE: usize = 3;
+pub const READ_CHUNK_SIZE: u64 = 64 * 1024;
+const CONNECTION_POOL_SIZE: u64 = 3;
 
 /// A [`StoreDriver`] implementation that uses Redis as a backing store.
 #[derive(MetricsComponent)]
@@ -178,7 +178,7 @@ impl StoreDriver for RedisStore {
     async fn has_with_results(
         self: Pin<&Self>,
         keys: &[StoreKey<'_>],
-        results: &mut [Option<usize>],
+        results: &mut [Option<u64>],
     ) -> Result<(), Error> {
         // TODO(caass): Optimize for the case where `keys.len() == 1`
         let pipeline = self.client_pool.next().pipeline();
@@ -334,8 +334,8 @@ impl StoreDriver for RedisStore {
         self: Pin<&Self>,
         key: StoreKey<'_>,
         writer: &mut DropCloserWriteHalf,
-        offset: usize,
-        length: Option<usize>,
+        offset: u64,
+        length: Option<u64>,
     ) -> Result<(), Error> {
         // To follow RBE spec we need to consider any digest's with
         // zero size to be existing.
@@ -355,7 +355,7 @@ impl StoreDriver for RedisStore {
         // We want to read the data at the key from `offset` to `offset + length`.
         let data_start = offset;
         let data_end = data_start
-            .saturating_add(length.unwrap_or(isize::MAX as usize))
+            .saturating_add(length.unwrap_or(isize::MAX as u64))
             .saturating_sub(1);
 
         // And we don't ever want to read more than `READ_CHUNK_SIZE` bytes at a time, so we'll need to iterate.
